Encoders! (#9)
KEVIN: Closing thoughts. A lot of things made sense when I realized what RamseteCommand and associated WPILib path following things actually did -- given a trajectory, the path following generates a list of left and right velocities for the drivetrain to follow the trajectory. My confusion from the PID oscillation stemmed from the assumption that the PID loop in the command was for position, so an oscillation would surely cause the robot to move back and forth; it never did because the PID loop was for velocity instead (this is why kI and kD were set to 0 in the example; accelerating to a target velocity is a use case for just a P controller). Oscillations due to a high kP were between 1.9m/s and 2.1m/s and not 1.9m and 2.1m so there was no back and forth driving. Furthermore, the confusion on why the robot stopped before the end position when constants were misconfigured can also be explained by realizing that the list of velocities probably ended, with the constants being incorrect leading to the robot not following the velocities properly and ending in an incorrect spot.

A lot of these confusions could probably have been resolved sooner had I read the docs for RamseteController a little more carefully.

* Initial testing

* Add wheel circumference to hopefully calculate distance

* Add position getting/setting, fix constants

* Slow drive speed, zero sensors

* Start odometry threading

* Odometry package and class

This commit will be overridden by using WPI's builtin odometry classes, but I am pushing it to keep it in history.

* Use WPI odometry class

* Small changes

* navX AHRS

* Merge left and right sensor methods to use odometry class

* Change `new Pose2d()` to `getRobotPosition()`

* Measure the constant!

This can be tested and tuned over time

* Bind A button to zero position

* Change constant to inches, remeasure

* Add better documentation

* Remove joysticks from encoder branch

* Readd odometry thread

Wasn't able to test this that much because the drivetrain became possessed by ghosts, but can test next week on the actual drivetrain.

* Talon -> Spark Max

* Implement preliminary path following command

See comment in FollowPathCommand

* Remove states from `tankSubsystem`

* Inches -> meters

* Add secondary constructor to FollowPathCommand

* Fix turning direction; remove CAN ID 0

* Edit to functional drivetrain state

* Measure some constants

Prepared to do https://docs.wpilib.org/en/stable/docs/software/wpilib-tools/robot-characterization/characterization-routine.html#running-the-characterization-routine and measure the rest on Sunday, then perhaps test finally!

* Measure more constants!

These will probably have to be remeasured with the Robot Characterization tool. When testing pathfinding with these, the robot immediately swerves sharply to the left (which it shouldn't be doing). It might have something to do with the right encoder readings being inverted during characterization (left reports 4 meters, right reports -4), but I've tried setting various options in the config to try and fix that to no avail. Will definitely look into it further at some point.

* Revert constants, invert gyro, invert left motors

* Fix inverted gyro error in resetPosition

* New constants!

Recharacterized using process described in https://www.chiefdelphi.com/t/the-future-of-pathweaver/393014/82. `Kp` might need to be tuned, as does `ENCODER_TICKS_TO_METERS` most likely.

* Update WPILib and vendor packages to 2022

Also push experimental changes to PID constants and auton testing, which still doesn't quite work

* Add preliminary shuffleboard code

* Move odometry to `tankSubsystem`

With the new REV version, CANSparkMaxes with the same can ID cannot be created (it throws errors which kill the robot program)

* Measure new constants with sysid

Hopefully these ones work better!

* `Field2d`, new constants

* Cleanup, remove elevator subsystem

* Change angle entry to use a gyro complex widget

Co-authored-by: pinkbluesky <pinkbluesky@users.noreply.github.com>