Use an abstract class for `GRTSubsystem`
This allows easier shared logic between subsystems (an interface would require each subsystem to implement their own versions of every method as it does not have access to fields). The original idea was to use this for climb cleanup as well as brownout, but basically anything that needs to be shared across subsystems can exist here.